+++ eemont/.ipynb_checkpoints/imagecollection-checkpoint.py
@@ -510,7 +510,7 @@ def spectralIndices(
     p=2.0,
     c=1.0,
 ):
-    """Computes one or more spectral indices (indices are added as bands) for an image collection.
+    """Computes one or more spectral indices (indices are added as bands) for an image collection from the Awesome List of Spectral Indices.
 
     Tip
     ----------